Misc cleanup.
