+++ moq-web/src/publish/frontend.rs
@@ -68,19 +68,20 @@ impl Publish {
 	pub fn set_volume(&mut self, volume: f64) {
 		self.controls.volume.set(volume);
 	}
-
-	#[wasm_bindgen(getter)]
-	pub fn closed(&self) -> bool {
-		self.controls.close.get()
-	}
-
-	pub fn close(&mut self) {
-		self.controls.close.set(true);
-	}
 }
 
 impl Default for Publish {
 	fn default() -> Self {
 		Self::new()
 	}
+}
+
+#[derive(Debug, Default, Copy, Clone)]
+#[wasm_bindgen]
+pub enum PublishState {
+	#[default]
+	Idle,
+	Connecting,
+	Connected,
+	Closed,
 }
