+++ moq-web/src/publish/mod.rs
@@ -1,12 +1,8 @@
 mod backend;
-mod controls;
 mod frontend;
-mod status;
 mod video;
 
 pub use frontend::*;
 
 use backend::*;
-use controls::*;
-use status::*;
 use video::*;
