--- moq-web/src/watch/controls.rs
@@ -1,11 +1,0 @@
-use baton::Baton;
-use url::Url;
-
-#[derive(Debug, Default, Baton)]
-pub struct Controls {
-	pub url: Option<Url>,
-	pub paused: bool,
-	pub volume: f64,
-	pub canvas: Option<web_sys::OffscreenCanvas>,
-	pub close: bool,
-}
