+++ moq-web/src/watch/mod.rs
@@ -1,14 +1,9 @@
 mod backend;
-mod controls;
 mod frontend;
 mod renderer;
-mod status;
 mod video;
 
-pub use frontend::*;
-
 use backend::*;
-use controls::*;
+pub use frontend::*;
 use renderer::*;
-use status::*;
 use video::*;
