--- moq-web/src/watch/status.rs
@@ -1,10 +1,0 @@
-use baton::Baton;
-
-use super::WatchState;
-use crate::Error;
-
-#[derive(Debug, Default, Baton)]
-pub struct Status {
-	pub state: WatchState,
-	pub error: Option<Error>,
-}
